django_project/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-model-utils==4.0.0
django-phonenumber-field==4.0.0
idna==2.8
Pillow==6.1.0
pytz==2019.2
requests==2.22.0
urllib3==1.24.2
